# rendering optimization and make clone method public