`upsteam/master` branch:

> $ go test -bench .
> PASS
> BenchmarkTableRenderTerminal-8         1        1971373200 ns/op
> BenchmarkTableRenderMarkdown-8         1        1904377500 ns/op
> BenchmarkTableRenderHTML-8             1        2220570800 ns/op
> ok      github.com/horpto/termtables    6.703s

PR branch:

> $ go test -bench .
> PASS
> BenchmarkTableRenderTerminal-8         5         332652980 ns/op
> BenchmarkTableRenderMarkdown-8         3         386607766 ns/op
> BenchmarkTableRenderHTML-8             2         688477350 ns/op
> ok      github.com/horpto/termtables    16.314s